Initial commit in own repo
This program has been moved to its own repo from a private repo of my research
lab. This is why it starts at v0.2.0.
The main motivation is so that consumers of stripped BAM files will be able
to understand why fields are missing, as they can see from the @PG headers that
`samstrip` has been used.